+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [2.7.0]
+
+- Individual Lab/Assistant Import & Export
+- Directly open workflows by clicking on tree items
+
 ## [2.6.2]
 
 - Fix config bug on swapping configuration contexts for import/export